OpenAI_sdk_Chatbot_Agent/Chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hassan", email = "hr8581492@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.11",
]

[project.scripts]
chatbot = "chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
